requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
SQLAlchemy==1.4.54
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-openlineage==2.2.0
pyspark==3.5.2